+++ Modules/Intent.Modules.AspNetCore.Controllers/release-notes.md
@@ -1,3 +1,7 @@
+### Version 6.0.9
+
+- Fixed: Issue where Policies where not added to controllers.
+
 ### Version 6.0.8
 
 - Fixed: Due to the nature of how GetTypeName resolves namespaces there are cases where ambiguous references still exist and causes compilation errors, this fix forces to re-evaluate a lot of types in this module.